requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
pydantic-settings==2.2.1
python-multipart==0.0.9
opencv-python==4.9.0.80
pdf2image==1.17.0
pytesseract==0.3.10
Pillow==10.2.0
